requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
PyPDF2
python-docx
python-dotenv
langchain
sentence-transformers
faiss-cpu